chore(app): update root components
